impl favor auto-insert
Former-commit-id: bab9558c12baddbb6210c05254f6327f00705526 [formerly 7a21ca4ae10946312f42c33cf0f497e6811866c8]
Former-commit-id: 74ae09432adc9e8593ff61837ec255e482a8c9cf

+++ backend/src/main/java/com/project/model/user/FavorEntity.java
@@ -31,7 +31,7 @@ public class FavorEntity {
     private String uid;
 
     @Column(name = "frid")
-    private int frid;
+    private long frid;
 
     // @OneToOne
     // @JoinTable(name = "ffff", //조인테이블명
